optimize code

+++ server/VueBlogNeteaseCloudMusicApiUploadServer/src/main/java/cn/smilex/vueblog/handler/Distribution.java
@@ -4,8 +4,8 @@ import cn.smilex.req.HttpRequest;
 import cn.smilex.req.HttpResponse;
 import cn.smilex.req.Requests;
 import cn.smilex.vueblog.Application;
-import cn.smilex.vueblog.config.MessageCode;
 import cn.smilex.vueblog.protocol.Message;
+import cn.smilex.vueblog.util.BucketUtil;
 import cn.smilex.vueblog.util.MessageUtil;
 import com.upyun.RestManager;
 import com.upyun.UpYunUtils;
@@ -34,31 +34,36 @@ public class Distribution {
 
                 Thread.ofVirtual()
                         .start(() -> {
-                            HttpResponse httpResponse = Requests.requests
-                                    .request(
-                                            HttpRequest.build()
-                                                    .setUrl(url)
-                                                    .setMethod(Requests.REQUEST_METHOD.GET)
-                                                    .setEnableDataByte(true)
-                                    );
+                            if (BucketUtil.getFileExists(filePath)) {
+                                MessageUtil.buildAndSendResponseMessage(
+                                        ctx.channel(),
+                                        filePath,
+                                        musicId
+                                );
+                            } else {
+                                HttpResponse httpResponse = Requests.requests
+                                        .request(
+                                                HttpRequest.build()
+                                                        .setUrl(url)
+                                                        .setMethod(Requests.REQUEST_METHOD.GET)
+                                                        .setEnableDataByte(true)
+                                        );
 
-                            var params = new HashMap<String, String>(1);
-                            params.put(RestManager.PARAMS.CONTENT_MD5.getValue(), UpYunUtils.md5(httpResponse.getDataByte()));
-                            try {
-                                var response = Application.REST_MANAGER
-                                        .writeFile(filePath, httpResponse.getDataByte(), params);
-                                if (response.isSuccessful()) {
-                                    var responseMessageContent = new HashMap<String, Object>(1);
-                                    responseMessageContent.put("url", Application.COMMON_CONFIG.getUrl() + filePath);
-                                    responseMessageContent.put("musicId", musicId);
-                                    MessageUtil.buildAndMessageMessage(
-                                            ctx.channel(),
-                                            MessageCode.RESPONSE_UPLOAD_RESULT,
-                                            responseMessageContent
-                                    );
+                                var params = new HashMap<String, String>(1);
+                                params.put(RestManager.PARAMS.CONTENT_MD5.getValue(), UpYunUtils.md5(httpResponse.getDataByte()));
+                                try {
+                                    var response = Application.REST_MANAGER
+                                            .writeFile(filePath, httpResponse.getDataByte(), params);
+                                    if (response.isSuccessful()) {
+                                        MessageUtil.buildAndSendResponseMessage(
+                                                ctx.channel(),
+                                                filePath,
+                                                musicId
+                                        );
+                                    }
+                                } catch (Exception e) {
+                                    log.error("", e);
                                 }
-                            } catch (Exception e) {
-                                log.error("", e);
                             }
                         });
                 break;
